restructuring for building/installing python nodes
